Creation of Chatterbox
This is chatterbox. It is a messaging app and you get to create your
own server. ikhaliq15 and nt developed this application in java and
exported them into runnable jar files. This software is totally free
and you can edit the source code in anyway you would like as long as
you give us credit.

Thanks,
     The Interweb Organization